requirements.txt
pinecone>=3.0.0
sentence-transformers>=2.2.2
requests>=2.31.0
python-dotenv>=1.0.0
groq>=0.4.0
